chore(cli): move lazy.ts and errors.ts to wiring/
These are CLI framework infrastructure files that belong alongside
createCliPlugin and CliInitializer rather than in the general lib/
directory.
