+++ packages/cli/src/modules/info/index.ts
@@ -15,7 +15,7 @@
  */
 import yargs from 'yargs';
 import { createCliPlugin } from '../../wiring/factory';
-import { lazy } from '../../lib/lazy';
+import { lazy } from '../../wiring/lazy';
 
 export default createCliPlugin({
   pluginId: 'info',
