+++ packages/cli/src/modules/translations/index.ts
@@ -15,7 +15,7 @@
  */
 import yargs from 'yargs';
 import { createCliPlugin } from '../../wiring/factory';
-import { lazy } from '../../lib/lazy';
+import { lazy } from '../../wiring/lazy';
 import { DEFAULT_MESSAGE_PATTERN } from './lib/messageFilePath';
 
 export default createCliPlugin({
