+++ packages/cli/src/wiring/CliInitializer.ts
@@ -20,7 +20,7 @@ import { CommandRegistry } from './CommandRegistry';
 import { Command } from 'commander';
 import { version } from '../lib/version';
 import chalk from 'chalk';
-import { exitWithError } from '../lib/errors';
+import { exitWithError } from './errors';
 import { ForwardedError } from '@backstage/errors';
 import { isPromise } from 'node:util/types';
