+++ packages/cli/src/wiring/errors.ts
